refactor: rename constants directory
